Finish merge of eth module
